pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "panelizer"
version = "0.1.0"
description = "Panel-by-panel comic reader with human-in-the-loop detection"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.9.0",
  "Pillow>=10.4.0",
  "pymupdf>=1.24.0",
  "opencv-python-headless>=4.10.0",
  "numpy>=2.1.0",
  "click>=8.1.0",
]

[project.scripts]
panelizer = "panelizer.__main__:main"
panel-flow = "panelizer.__main__:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=6.0.0",
  "ruff>=0.6.0",
  "pyright>=1.1.385",
  "twine>=5.0.0",
]
ml = [
  "torch>=2.0.0",
  "ultralytics>=8.3.0",
  "huggingface_hub>=0.20.0",
]
ml-cpu = [
  "torch>=2.0.0,<2.7",
  "torchvision>=0.15.0,<0.22",
  "ultralytics>=8.3.0",
  "huggingface_hub>=0.20.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=6.0.0",
  "ruff>=0.6.0",
  "pyright>=1.1.385",
  "twine>=5.0.0",
]
ml = [
  "torch>=2.0.0",
  "ultralytics>=8.3.0",
  "huggingface_hub>=0.20.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/panelizer"]
include = ["src/panelizer/preview/static/**"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --cov=panelizer --cov-report=term-missing --cov-fail-under=80"

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["panelizer"]

[tool.pyright]
include = ["src", "tests"]
exclude = [
  "**/__pycache__",
  ".git",
  ".venv",
  "build",
  "dist",
  ".pytest_cache",
  ".ruff_cache",
]
pythonVersion = "3.11"
typeCheckingMode = "basic"
reportMissingImports = "warning"
reportMissingModuleSource = "warning"
venvPath = "."
venv = ".venv"

# uv configuration for CPU-only PyTorch
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true  # Only use when explicitly requested

[tool.uv.sources]
# When ml-cpu extra is used, get torch/torchvision from CPU index
torch = [
  { index = "pytorch-cpu", extra = "ml-cpu" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "ml-cpu" },
]
